drm/sun4i: Add TCON TOP driver
TCON TOP is special unit, which is responsible for determining
relationships between mixers and TCONs. Additionally, it also takes care
for determining which TCON should be source to HDMI. Last but not least,
it contains additional bus gates for tcon-tv and dsi units.

Something similar also existed in previous SoCs, except that it was part
of first TCON.

Signed-off-by: Jernej Skrabec <jernej.skrabec@siol.net>